requirements.txt
mysql-connector-python
python-dotenv
sqlalchemy
smartapi-python
yfinance
logzero
websocket
pyotp

